Set up the mitmproxy config as a symlink
This allows the mitmproxy configuration directory to be saved as a Docker volume so that its certificates are not regenerated every time the container is run. If the config file is directly included in the volume, then new changes to it will not be copied over during the image build because the volume mount (which is using the old config version) overwrites the whole directory. Using a symlink allows modifying the config without modifying anything in the configuration directory.